Refactor tests test/code and test/content
