+++ test/code/runners/typeql_runner.py
@@ -1,8 +1,8 @@
 from typedb.driver import TypeDB, Driver, TransactionType, Credentials, DriverOptions
-from test.runners.base_runner import BaseRunner, TEST_CONFIG_KEY_RESET
+from test.code.runners.base_runner import BaseRunner, TEST_CONFIG_KEY_RESET
 from enum import Enum
 from typing import List, Dict, Tuple, Union
-from test.parser.parser import ParsedTest
+from test.code.parser.parser import ParsedTest
 import logging
 logger = logging.getLogger('main')
 # To see debug log, set logging level to debug in main.py
