+++ test/code/tests/rust_test.adoc
