+++ test/content/mock/academy/antora.yml
